pyproject.toml
[project]
name = "repositorio1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.8",
    "langchain>=0.3.19",
    "langchain-community>=0.3.18",
    "openai>=1.64.0",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.0",
]
